Remove i2c0, route buttons and OLED
i2c isn't used on the board, MIO pins 50 and 51 are connected to
the push buttons. Remove i2c0 for this to work correctly.

Route GPIO pins through logic so that the push buttons are connected
to GPIO lines. Idem for the OLED connections.

The XPS tool is rather fond of changing whitespace in files for
no apparent reason, so I did my best to get rid of those.