Replace FPGA i2c periferal with "hard" PS I2C
The chip has I2C hardware, so use it to save on resources and power.
Removed the iic periferal from the design.
Activated the i2c0 controller in the Zynq PS and connect to EMIO.
Wire the i2c0 controller lines to the i2c mixer in programmable logic.

This requires the PS IIC driver in the kernel, and an adapted device
tree to use the new devices.